review changes: UserCreationTests now with BaseTest and overridden setUp for Registration
