+++ src/test/java/ru/praktikum/diplom/UserChangeDataTests.java
@@ -1,4 +1,4 @@
-package ru.praktukum.diplom;
+package ru.praktikum.diplom;
 
 import io.qameta.allure.Description;
 import io.qameta.allure.junit4.DisplayName;
